[Fix] 반환 타입 변경

+++ coachcoach/catalog/src/main/java/com/coachcoach/catalog/dto/response/HomeMenusResponse.java
@@ -3,7 +3,7 @@ package com.coachcoach.catalog.dto.response;
 import java.math.BigDecimal;
 
 public record HomeMenusResponse(
-        int numOfDangerMenus,
+        Long numOfDangerMenus,
         AvgCostRate avgCostRate,
         BigDecimal avgMarginRate
 ) {
